debug

+++ batchgen/models/deepseek/deepseekv3/modeling_deepseek_v3.py
@@ -865,7 +865,10 @@ class DeepseekV3MoE_Prefill(nn.Module):
 		if not self.training:
 			y = self.moe_infer(hidden_states, topk_idx, topk_weight).view(*orig_shape)
 		if self.config.n_shared_experts is not None:
-			y = y + self.shared_experts(identity)
+			# y = y + self.shared_experts(identity)
+			# accumulate in fp32
+			y = y.to(torch.float32) + self.shared_experts(identity).to(torch.float32)
+			y = y.type(identity.dtype)
 		return y
 
 	@torch.no_grad()
@@ -1227,7 +1230,13 @@ class DeepseekV3MoE_Decoding_FP8(nn.Module):
 		
 		out = self.moe_infer_allgather_allreduce_opt(hidden_states)
 		# out = self.moe_infer_alltoall(hidden_states)
-		out = out + self.shared_experts(identity)
+		# out = out + self.shared_experts(identity)
+		# accumulate with fp32
+		out = out.float()
+		out = out + self.shared_experts(identity).float()
+		out = out.to(hidden_states.dtype)
+
+
 		return out.view(*orig_shape)
 	
 	@torch.inference_mode()
